♻️ Refactor add videos route into FSD slices
